Refactor code to use common libs
